scrapper/pyproject.toml
[tool.poetry]
name = "hn_scrape"
version = "0.1.0"
description = "Scraping of hackernews + Build the DB"
authors = ["Omar MOHAMMED"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.6.2"
pymongo = "^4.6.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
